delete test_grid_sampler_3d.py
