+++ paddle/fluid/operators/grid_sampler_op.cc
@@ -61,8 +61,7 @@ class GridSampleOpMaker : public framework::OpProtoAndCheckerMaker {
         "(Tensor) The input grid of GridSampleOp generated by AffineGridOp, "
         "This is a 4-D tensor with shape of [N, H, W, 2] is the concatenation "
         "of x and y coordinates with shape [N, H, W] in last dimension or "
-        "        a 5-D tensor with shape of [N, D, H, W, 3] is the "
-        "concatenation "
+        "a 5-D tensor with shape of [N, D, H, W, 3] is the concatenation "
         "of depth, x and y coordinates with shape [N, D, H, W] in last "
         "dimension ");
     AddOutput("Output",
