+++ paddle/phi/kernels/gpu/grid_sample_kernel.cu
@@ -370,80 +370,66 @@ void GridSampleKernel(const Context& dev_ctx,
     VLOG(3) << "out dims: " << out->dims()[0] << "; " << out->dims()[1] << "; "
             << out->dims()[2] << "; " << out->dims()[3];
 
-    if (x.dims().size() == 4) {
-      const int n = grid.dims()[0];
-      const int out_h = grid.dims()[1];
-      const int out_w = grid.dims()[2];
-      const int c = x.dims()[1];
-      const int in_h = x.dims()[2];
-      const int in_w = x.dims()[3];
-      VLOG(3) << "n: " << n << "; c: " << c << "; out_h: " << out_h
-              << "; out_w: " << out_w;
+    int count = static_cast<int>(n * out_h * out_w);
+    auto cu_stream = dev_ctx.stream();
+    backends::gpu::GpuLaunchConfig config =
+        backends::gpu::GetGpuLaunchConfig1D(dev_ctx, count);
+    GridSampleCudaKernel<T>
+        <<<config.block_per_grid, config.thread_per_block, 0, cu_stream>>>(
+            count,
+            n,
+            c,
+            out_h,
+            out_w,
+            in_h,
+            in_w,
+            x.data<T>(),
+            grid.data<T>(),
+            output_data,
+            enum_mode,
+            enum_padding_mode,
+            align_corners);
+  } else {
+    const int n = grid.dims()[0];
+    const int out_d = grid.dims()[1];
+    const int out_h = grid.dims()[2];
+    const int out_w = grid.dims()[3];
+    const int c = x.dims()[1];
+    const int in_d = x.dims()[2];
+    const int in_h = x.dims()[3];
+    const int in_w = x.dims()[4];
 
-      auto* output_data = dev_ctx.template Alloc<T>(out);
-      VLOG(3) << "out dims: " << out->dims()[0] << "; " << out->dims()[1]
-              << "; " << out->dims()[2] << "; " << out->dims()[3];
+    VLOG(3) << "n: " << n << "; c: " << c << "; out_d: " << out_d
+            << "; out_h: " << out_h << "; out_w: " << out_w;
 
-      int count = static_cast<int>(n * out_h * out_w);
-      auto cu_stream = dev_ctx.stream();
-      backends::gpu::GpuLaunchConfig config =
-          backends::gpu::GetGpuLaunchConfig1D(dev_ctx, count);
-      GridSampleCudaKernel<T>
-          <<<config.block_per_grid, config.thread_per_block, 0, cu_stream>>>(
-              count,
-              n,
-              c,
-              out_h,
-              out_w,
-              in_h,
-              in_w,
-              x.data<T>(),
-              grid.data<T>(),
-              output_data,
-              enum_mode,
-              enum_padding_mode,
-              align_corners);
-    } else {
-      const int n = grid.dims()[0];
-      const int out_d = grid.dims()[1];
-      const int out_h = grid.dims()[2];
-      const int out_w = grid.dims()[3];
-      const int c = x.dims()[1];
-      const int in_d = x.dims()[2];
-      const int in_h = x.dims()[3];
-      const int in_w = x.dims()[4];
+    auto* output_data = dev_ctx.template Alloc<T>(out);
+    VLOG(3) << "out dims: " << out->dims()[0] << "; " << out->dims()[1] << "; "
+            << out->dims()[2] << "; " << out->dims()[3] << "; "
+            << out->dims()[4];
 
-      VLOG(3) << "n: " << n << "; c: " << c << "; out_d: " << out_d
-              << "; out_h: " << out_h << "; out_w: " << out_w;
-
-      auto* output_data = dev_ctx.template Alloc<T>(out);
-      VLOG(3) << "out dims: " << out->dims()[0] << "; " << out->dims()[1]
-              << "; " << out->dims()[2] << "; " << out->dims()[3] << "; "
-              << out->dims()[4];
-
-      int count = static_cast<int>(n * out_d * out_h * out_w);
-      auto cu_stream = dev_ctx.stream();
-      backends::gpu::GpuLaunchConfig config =
-          backends::gpu::GetGpuLaunchConfig1D(dev_ctx, count);
-      GridSample3DCudaKernel<T>
-          <<<config.block_per_grid, config.thread_per_block, 0, cu_stream>>>(
-              count,
-              c,
-              out_d,
-              out_h,
-              out_w,
-              in_d,
-              in_h,
-              in_w,
-              x.data<T>(),
-              grid.data<T>(),
-              output_data,
-              enum_mode,
-              enum_padding_mode,
-              align_corners);
-    }
+    int count = static_cast<int>(n * out_d * out_h * out_w);
+    auto cu_stream = dev_ctx.stream();
+    backends::gpu::GpuLaunchConfig config =
+        backends::gpu::GetGpuLaunchConfig1D(dev_ctx, count);
+    GridSample3DCudaKernel<T>
+        <<<config.block_per_grid, config.thread_per_block, 0, cu_stream>>>(
+            count,
+            c,
+            out_d,
+            out_h,
+            out_w,
+            in_d,
+            in_h,
+            in_w,
+            x.data<T>(),
+            grid.data<T>(),
+            output_data,
+            enum_mode,
+            enum_padding_mode,
+            align_corners);
   }
 }
+
 }  // namespace phi
 
 PD_REGISTER_KERNEL(
